Network should now be fully working.

+++ Client/include/ClientInventory.h
@@ -7,21 +7,21 @@ class Item;
 class ClientInventory {
 private:
     QList<Item*> items;
-    int max_weight;
-    int current_weight;
+    quint16 max_weight;
+    quint16 current_weight;
     QList<Item*> equipped;
 
 public:
-    ClientInventory(int);
+    ClientInventory(quint16);
     ClientInventory(const ClientInventory&);
     QList<Item*> get_items();
     void add_item(Item*);
     void remove_item(Item*);
     bool has_item(Item*) const;
     void set_max_weight(Item*);
-    int get_weight() const;
-    int get_max_weight() const;
-    void set_max_weight(int);
+    quint16 get_weight() const;
+    quint16 get_max_weight() const;
+    void set_max_weight(quint16);
     QList<Item*> get_equipped() const;
     void equip(Item*);
     void unequip(Item*);
